Rename base.py to more fitting base_schema.py.
